refactor(ntc-oracle-node): move auth_data functions into handler
