+++ rust-workspace/crates/ntc-oracle-node/Cargo.toml
@@ -8,14 +8,15 @@ edition = "2021"
 [dependencies]
 algonaut = { git = "https://github.com/manuelmauro/algonaut", rev = "30a251e438df9bb7af8b3aafc53bb9945a74c963" }
 algonaut_client = { git = "https://github.com/manuelmauro/algonaut", rev = "30a251e438df9bb7af8b3aafc53bb9945a74c963" }
-axum = "0.5.6"
-tokio = { version = "1", features = ["full"] }
-config = "0.13.1"
+axum = "0.5"
+base64 = "0.13"
+config = "0.13"
+ed25519 = { version = "1.5", features = ["pkcs8"] }
+ring-compat = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_derive = "1.0"
-base64 = "0.13.0"
 serde_json = "1.0"
-ed25519 = {version="1.5", features = ["pkcs8"]}
-ring-compat = "0.4.1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+tower-http = { version = "0.3", features = ["cors"] }
 
 env-var-helpers = { git = "https://github.com/PiDelport/rust-env-var-helpers" }
