+++ rust-workspace/crates/ntc-oracle-node/src/handlers/mod.rs
@@ -1,0 +1,3 @@
+//! HTTP handlers.
+
+pub(crate) mod auth_data;
