+++ rust-workspace/crates/ntc-oracle-node/src/signer.rs
@@ -1,8 +1,8 @@
 use ed25519::signature::{Signer, Verifier};
 
-use crate::AuthData;
+use crate::handlers::auth_data::AuthData;
 
-pub struct AuthDataSigner<S>
+pub(crate) struct AuthDataSigner<S>
 where
     S: Signer<ed25519::Signature>,
 {
@@ -13,7 +13,7 @@ impl<S> AuthDataSigner<S>
 where
     S: Signer<ed25519::Signature>,
 {
-    pub fn sign(&self, auth_data: &AuthData) -> ed25519::Signature {
+    pub(crate) fn sign(&self, auth_data: &AuthData) -> ed25519::Signature {
         // NOTE: use `try_sign` if you'd like to be able to handle
         // errors from external signing services/devices (e.g. HSM/KMS)
         // <https://docs.rs/signature/latest/signature/trait.Signer.html#tymethod.try_sign>
@@ -22,7 +22,7 @@ where
     }
 }
 #[allow(dead_code)]
-pub struct AuthDataVerifier<V> {
+pub(crate) struct AuthDataVerifier<V> {
     pub verify_key: V,
 }
 
@@ -31,7 +31,7 @@ where
     V: Verifier<ed25519::Signature>,
 {
     #[allow(dead_code)]
-    pub fn verify(
+    pub(crate) fn verify(
         &self,
         auth_data: &AuthData,
         signature: &ed25519::Signature,
